+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.0] - 2018-10-25
+### Added
+ - Allows to remove REST API Really Simple Discovery (RSD) endpoint.
+ - Allows to remove `<link>` to the Really Simple Discovery service endpoint on front-end pages.
+ - Allows to remove `X-Pingback` HTTP Header on front-end pages.
+
+### Changed
+ - Updated translations and screenshots.
+ - Documentation improvements.
+
 ## [1.0.1] - 2018-10-18
 ### Added
  - Allows to remove REST and Discovery oEmbed `<link>` tags on front-end pages.